fix: restructure positions and transactions
