+++ src/pages/Positions/index.tsx
@@ -1,9 +1,9 @@
-import React, { useContext, useState } from 'react'
+import React, { useContext } from 'react'
 import { Link } from 'react-router-dom'
 import { Helmet } from 'react-helmet'
 import { ThemeContext } from 'styled-components'
 import { getExplorerLink, shortenAddress } from '../../utils'
-import TransactionHistory from './TransactionHistory'
+import LiquidityPositions from './LiquidityPositions'
 import { ChevronLeft, User, Copy, ExternalLink } from 'react-feather'
 import { Button, Dots } from 'components'
 import { useActiveWeb3React } from '../../hooks/useActiveWeb3React'
@@ -19,7 +19,7 @@ const mock = {
     ]
 }
 
-export default function Transactions() {
+export default function Positions() {
     const theme = useContext(ThemeContext)
     const { account, chainId } = useActiveWeb3React()
     const userEthBalance = useETHBalances(account ? [account] : [])?.[account ?? '']
@@ -27,7 +27,7 @@ export default function Transactions() {
     return (
         <>
             <Helmet>
-                <title>Transactions | Sushi</title>
+                <title>Positions | Sushi</title>
             </Helmet>
 
             {/* <div className="w-full max-w-2xl">
@@ -85,17 +85,17 @@ export default function Transactions() {
 
             <div className="bg-dark-900 w-full max-w-2xl rounded p-4">
                 <div className="w-auto flex justify-between items-center rounded bg-dark-800 p-0.5 whitespace-nowrap text-sm font-bold cursor-pointer select-none pointer-events-auto mt-3 mb-6">
-                    <Link to={'/pool'} className={`w-3/6 p-3 text-center rounded-lg text-secondary`}>
-                        Liquidity Positions
-                    </Link>
                     <Link
-                        to={'/transactions'}
+                        to={'/pool'}
                         className={`w-3/6 p-3 text-center rounded-lg text-primary text-bold bg-dark-900 `}
                     >
+                        Liquidity Positions
+                    </Link>
+                    <Link to={'/transactions'} className={`w-3/6 p-3 text-center rounded-lg text-secondary`}>
                         Transaction History
                     </Link>
                 </div>
-                <TransactionHistory transactions={mock.transactions} />
+                <LiquidityPositions />
             </div>
         </>
     )
