+++ src/routes.tsx
@@ -35,7 +35,8 @@ import Tools from './pages/Tools'
 import Vesting from './pages/Vesting'
 import MasterChefV1 from './pages/Yield/masterchefv1'
 import MiniChefV2 from './pages/Yield/minichefv2'
-import Transactions from 'pages/Transactions'
+import Positions from './pages/Positions'
+import Transactions from './pages/Transactions'
 
 function Routes(): JSX.Element {
     const { chainId } = useActiveWeb3React()
@@ -85,7 +86,7 @@ function Routes(): JSX.Element {
             <Route exact strict path="/swap/:outputCurrency" component={RedirectToSwap} />
             <Route exact strict path="/send" component={RedirectPathToSwapOnly} />
             <Route exact strict path="/find" component={PoolFinder} />
-            <Route exact strict path="/pool" component={Pool} />
+            <Route exact strict path="/pool" component={Positions} />
             <Route exact strict path="/transactions" component={Transactions} />
             <Route exact strict path="/create" component={RedirectToAddLiquidity} />
             <Route exact path="/add" component={AddLiquidity} />
